Add semantic cache files and new DTO for report concepts
- Created multiple JSON files in the semantic cache to represent various domain entities, repositories, and their relationships, including:
  - IProyectoRepositorio
  - Web Program
  - ServicioConfiguracion
  - FinanzasEndpoints
  - ReportesEndpoints
  - IPreferenciasRepositorio
  - TemaCambiado event
  - PreferenciasRepositorio
  - PreferenciasConfiguracion

- Introduced a new DTO: ReporteConConceptosDto to encapsulate report details along with associated concepts.

+++ src/FinanKore.Application/Proyecto/Comandos/ActualizarConceptoReporteManejador.cs
@@ -28,7 +28,7 @@ public sealed class ActualizarConceptoReporteManejador(
             concepto.Id,
             concepto.Nombre,
             concepto.Valor,
-            concepto.Tipo,
+            (int)concepto.Tipo,
             concepto.ReporteId,
             concepto.CategoriaId,
             concepto.FechaCreacion);
