+++ src/FinanKore.Application/Proyecto/Comandos/CrearConceptoReporteManejador.cs
@@ -25,7 +25,7 @@ public sealed class CrearConceptoReporteManejador(
             concepto.Id,
             concepto.Nombre,
             concepto.Valor,
-            concepto.Tipo,
+            (int)concepto.Tipo,
             concepto.ReporteId,
             concepto.CategoriaId,
             concepto.FechaCreacion);
